:recycle: (config-eslint) files renamed as mjs
