+++ packages/config-eslint/package.json
@@ -5,17 +5,16 @@
   "type": "module",
   "devDependencies": {
     "@antfu/eslint-config": "^3.12.0",
-    "@eslint/eslintrc": "^3",
+    "@next/eslint-plugin-next": "^15.1.5",
     "eslint": "^9.17.0",
-    "eslint-config-next": "15.1.5",
     "eslint-plugin-drizzle": "^0.2.3",
     "eslint-plugin-format": "^0.1.3",
     "eslint-plugin-turbo": "^2.3.3"
   },
   "exports": {
-    ".": "./eslint.config.js",
-    "./create-config": "./create-config.js",
-    "./hono": "./hono.js",
-    "./next": "./next.js"
+    ".": "./eslint.config.mjs",
+    "./create-config": "./create-config.mjs",
+    "./hono": "./hono.mjs",
+    "./next": "./next.mjs"
   }
 }
